refactor: 웹 변환 작업대 feature 구조 분리
- 변환 작업대 상태와 API 호출을 useConverter 훅과 lib 유틸로 분리
- 컨트롤 데크와 미리보기 작업 영역 UI를 feature 하위 컴포넌트로 이동
- 웹 앱 구조 변경 내용을 architecture와 memory 문서에 반영
