+++ apps/web/src/features/converter/lib/image.ts
@@ -1,0 +1,15 @@
+import type { ImageMeta } from "../types";
+
+export function readImageMeta(url: string): Promise<ImageMeta> {
+  return new Promise((resolve, reject) => {
+    const image = new window.Image();
+    image.onload = () => {
+      resolve({
+        width: image.naturalWidth,
+        height: image.naturalHeight,
+      });
+    };
+    image.onerror = reject;
+    image.src = url;
+  });
+}
